Improve stitch speed on larger datasets
Previously, parsing taken so long it was worth parsing just once and
retaining every parsed map in memory.

Now that parsing is fast, memory allocation has become the bottleneck
and it's worth just having only the current map in memory.

This is probably a slowdown on small datasets where the required memory
can be acquired quickly, but in those cases, hey, it's not long anyway.